Create Static Initializer Block
There are two lines of input. The first line contains : B the breadth of the parallelogram. The next line contains : H the height of the parallelogram.
You should read the variables from the standard input.
If  B<=0 or H<=0  , the output should be "java.lang.Exception: Breadth and height must be positive" without quotes.